backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.115.0
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic==2.5.1
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis and Celery
redis==5.0.1
celery==5.3.4
flower==2.0.1

# Authentication and Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.1.2
passlib[bcrypt]==1.7.4

# Google OAuth
authlib==1.2.1
httpx==0.28.1

# Stripe Integration
stripe==9.12.0

# File Processing
pandas==2.1.3
openpyxl==3.1.2
chardet==5.2.0
python-magic==0.4.27

# Gemini AI API
google-genai==1.30.0

# Email
resend==0.6.0

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
python-dotenv==1.0.0
email-validator==2.1.0
aiohttp>=3.9.0
